requirements.txt
email_validator
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Flask[async]>=2.0
google-generativeai
google-adk
langchain
langchain_google_genai
langchain_community
faiss-cpu
gunicorn>=20.0
PyPDF2
Werkzeug